refactoring - mooved notifications table to commons
