+++ packages/pn-pa-webapp/src/api/notifications/Notifications.api.ts
@@ -1,9 +1,7 @@
-import { formatDate } from '@pagopa-pn/pn-commons';
+import { formatDate, GetNotificationsParams, GetNotificationsResponse } from '@pagopa-pn/pn-commons';
 
 import { apiClient } from '../axios';
-import { GetNotificationsParams, GetNotificationsResponse } from '../../redux/dashboard/types';
 import { LegalFactId, NotificationDetail } from '../../redux/notification/types';
-
 
 export const NotificationsApi = {
   /**
