+++ packages/pn-pa-webapp/src/redux/dashboard/__test__/test-utils.ts
@@ -1,6 +1,4 @@
-import { formatDate, NotificationStatus } from "@pagopa-pn/pn-commons";
-
-import { GetNotificationsResponse } from "../types";
+import { formatDate, GetNotificationsResponse, NotificationStatus } from "@pagopa-pn/pn-commons";
 
 export const notificationsFromBe: GetNotificationsResponse = {
   result: [
