+++ packages/pn-pa-webapp/src/redux/dashboard/actions.ts
@@ -1,7 +1,7 @@
 import { createAction, createAsyncThunk } from "@reduxjs/toolkit";
+import { GetNotificationsParams, GetNotificationsResponse } from "@pagopa-pn/pn-commons";
 
 import { NotificationsApi } from "../../api/notifications/Notifications.api";
-import { GetNotificationsParams, GetNotificationsResponse } from "./types";
 
 export const getSentNotifications = createAsyncThunk<GetNotificationsResponse, GetNotificationsParams>
     ("getSentNotifications", async (params: GetNotificationsParams, { rejectWithValue }) => {
