+++ packages/pn-pa-webapp/src/redux/dashboard/reducers.ts
@@ -1,8 +1,8 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { today, tenYearsAgo } from "../../utils/date.utility";
+import { GetNotificationsParams, Notification } from "@pagopa-pn/pn-commons";
 
+import { today, tenYearsAgo } from "../../utils/date.utility";
 import { getSentNotifications, setPagination, setSorting, setNotificationFilters } from "./actions";
-import { GetNotificationsParams, Notification } from "./types";
 
 /* eslint-disable functional/immutable-data */
 const dashboardSlice = createSlice({
